[*] Swaps over implementation to using spawn instead of exec.
This is to protect against a injection vulnerability that
was identified by Alessio (d3lla).

Many thanks to Alessio for finding this and bringing it to
my attention.

A unit test that covers a specific example of this attack has
been added into the project to ensure that it errors instead
of executing the injected code.